Remove deprecated warnings for SKIP_VALIDATE
We deliberately use a deprecated API to test that it is not used. Ignore
warnings from using it in this test.

In PresentDisplayNoLayerStateChanges, remove the check for
SKIP_VALIDATE, removing the other warning. As written, the test would be
skipped without SKIP_VALIDATE support. But devices should *not* claim
Capability::SKIP_VALIDATE, meaning the test would always be skipped. The
docs state that the behavior previously indicated by SKIP_VALIDATE is
now the default, so we should always be running the test.

Bug: NA
Test: atest VtsHalGraphicsComposer3_TargetTest (no warnings)
Change-Id: Icc89a82b4850428c2f0940be0f78f5ebff1881b3