wifi: Enhance tombstone folder dumping behavior
In current implementation, files in the tombstone folder are only
archived when there is at least one active wifi interface. The
examination is unnecessary and it causes problems for us to get wifi
logs if reporter takes bugreport with STA/SAP off. In this change,
tombstone files will be archived regardless of the state of the wifi
interfaces.

Bug: 303429585
Test: dumpsys android.hardware.wifi.IWifi/default
Change-Id: I8a8b09f6473b2a1f93478b0393963e87b1fb9ad7