Make refreshTimestamp atomic.
Make refreshTimestamp atomic in VehiclePropertyStore so that we
are sure the property value will not be updated after we read
the value out, but before we put the value with the new timestamp back.

Otherwise there is a slight chance that a property update will be
overwritten by the timestamp refresh operation.

Test: atest FakeVehicleHardwareTest
Bug: 308552957
Change-Id: I789257b4c75d8a4d19edeec31d01dc23776233ec