wifi: Modified API for P2pProvisionDiscoveryCompleted event
Deprecated the existing onProvisionDiscoveryCompleted API
which is used to indicate the completion of a P2P provision
discovery request. Added a new modified API called
onProvisionDiscoveryCompletedEvent to include the P2P group interface
name. This is for framework to know which group owner interface
received the provision discovery request, when the Peer device tries
to join the existing P2P group owner.

Bug: 296063280
Test: vts test
Test: Ran P2P negotiated connection tests via Wi-Fi Direct settings
      window.
Change-Id: If349ed600f2a474d50c0b05d58331b5b8ba9e94d