Remove symlink to audio_effects_conf.xsd
A change is being made so that sandboxed actions need to add a
dependency on the targets of the symlinks they use. All xsd_config
modules use sandboxing by default.

In this case, the symlink isn't necessary, and we can just depend
on the audio_effects_conf_V2_0 genrule instead.

Bug: 307824623
Test: Presubmits
Change-Id: Ia3c2e8adced78b39881d8c0390c881c0d28ecde4