Change temp approximations to be accurate
We need to use a more accurate ratio of celsius to fahrenheit along
with better min/max temperature conversions so if a client sets the
value using standard conversion or using this table, they will produce
the same result. Also, we need to round to the closest increment instead
of truncating.

Bug: 305274504
Bug: 313720524
Test: atest CtsCarTestCases:CarPropertyManagerTest
Test: atest FakeVehicleHardwareTest
Change-Id: Ia4d53e1a904fb2b40f5ec428ab548895c8f307ed