Allow non-zero HCI interface numbers
[Description]
when driver re-register hci device, it may register hci 1
[Root Cause]
In the kernel mechanism, there may be a timing issue. when driver re-register hci device,kernel may not completely unregister hci 0.But aidl only open hci 0.
[Solution]
assign a value to the hci interface based on the device node found by AIDL

[Test Report]
Insert the bt chip into the usb 2.0 port, and then suspend the TV. The bt chip will be disconnected due to the USB power outage. And then quickly resume the TV，check whether the Bluetooth state is on. Suspend and resume for several times, check for bt state.

Change-Id: I57fd4a22c956e0cf2b4a516247c258493163e51a
BUG: 314149416