Fix timeout issue for bcradio worker thread
Accessed mIsTerminating inside the same lock as what is used in while
loop in worker thread class of broadcast radio HAL utils lib. This fixed
the race condition that mIsTerminating is accessed as true in threadLoop
while the destructor is setting mIsTerminating as false, which causes the
thread waits forever for lock after lock is released in the desctructor.

Bug: 314100017
Test: atest VtsHalBroadcastradioAidlTargetTest WorkerThreadTest
Change-Id: I885e1487ac39525fc2d1ee2134d24409264ca0fc