Secretkeeper: Disable a replay protection VTS
This test does not account for potentially missing Sk HAL. Ideally, the
test should pass on those devices too (since Sk is an optional HAL at
the moment).

This test uses should_panic expecting the client code to panic on
processing the message, but that code is never executed because the
backend is not present & the method doesn't panic & the test fails.

As a mitigation, #ignore the test.

Bug: 317416663
Test: VTS show "IGNORED"

Change-Id: I2f1bb45361943f22d8cadc929059c65cb3120f3f