Make remoteaccess HAL pass VTS.
The reference remote access HAL should still pass CTS even when
the grpc remote access server does not exist. The reference
remote access HAL now allows GRPC_SERVICE_ADDRESS not to be
defined. If it is not defined, it will not try to connect
a remote server and will work in a fake mode.

Test: VtsHalAutomotiveRemoteAccess_TargetTest with an without
grpc server running.
Bug: 277967402

Change-Id: I89509ac8f8ebe9a268d3a338d6e80c24e39dc512