Skip unsupported functinality tests
Instead of silently passing the tests when some functionality is
not supported, mark these tests as skipped. This way, when we are sure
that some functionality should be verified in a particular test run,
we are not being falsely reported that all have passed successfully
while in reality, there's an actuall issue in the particular functionality
detection mechanism.

Bug: 319090769
Test: atest VtsHalBluetoothAudioTargetTest
Change-Id: I8589ef14abfb5af641bc98949de1e1acc21efe1a