Fix tune byte alignment
When reading data from the socket, one byte is read during tuning
and remaining data is read during playback. Since the same streamer
is used for reading data, there is an offset of 1 when tuning is
performed before playback. The extra byte is stored in a buffer
and prepended to the playback buffer.

Bug: 288170590
Test: atest VtsHalTvTunerTargetTest
Change-Id: Ie5d112dbc3c3e3bbb0bb07e60d15ddc26cacaf8c