Add retry in setProp test.
Set property operation is async so getting the property value
immediately after setting the property might not return the new
value. This CL adds the logic to retry when we do not get the
expected values back.

This CL also skips the test case if getting property or setting
property returns unavailable.

Test: atest VtsHalAutomotiveVehicle_TargetTest
Bug: 322070490
Change-Id: If9349a097a92c51101c7b5f4bf807b610ab2cb0b