Update bump.py for Trunk Stable
The next year's compatbility matrix is added to a conditional statement
so it's only available on in-development release configurations.

The curent year's compatibility matrix is moved from the conditional
statement to always be added to all release configs.

The next year's compatibility matrix level is set to the next year's
level after copying the current matrix file.

Test: ./bump.py
Bug: 279809333

Change-Id: Id711ba79110c8775f715eddf37a9bf51b073ec91