bump.py remove Level.h dependency
We want to run bump.py during finalization. We don't want to have to
update Level.h and the VTS tests that early.
So this CL removes the dependency on Level.h and requires the
current/next letters of the API levels for the kernerl configs to be
passed as arguments.

Test: bump.py 202404 202505 v w
Bug: 279809333
Change-Id: If8f281eccf62d380949a5ea9f5d0d3bb2d7f19ab