Provide better documentation for CommandErrors
Some HWC implementations which do not support DISPLAY_DECORATION are
handling DISPLAY_DECORATION layers in VTS by reporting an error (as
required by the test) AND by proposing ChangedCompositionTypes. The
latter is unnecessary, as SF will never read the changed types. Update
the documentation to be more specific that this is unnecessary.

Bug: b/324977561
Bug: b/329119344
Bug: b/329017438
Bug: b/327322951
Bug: b/327090469
Test: GraphicsComposerAidlCommandTest#DisplayDecoration
Change-Id: I09c90e40818dc43b968c3e79e61e8b13a13e6344