Fix flaky TestWakeupClientServiceImplUnitTest.
Previously the test will block forever if buildAndStart returns
a nullptr if the test grpc server fails to start. This CL fixes
this issue by adding a timeout and stop the wait even if the return
value is nullptr. This CL also lets grpc randomly pick an available
port instead of hard coding the port to make test more stable.

Test: atest --host TestWakeupClientServiceImplUnitTest
Bug: 330115017
Change-Id: I1113f95f92a25254e850b7a2eec249d1101c0dfe