DisplayDecoration test: destroy layers per display
This appears to have been broken with
Ic3c0415612b1387671ab8efa0ec95ed727f6bf68, which attempts to destroy
layers for all displays using a single writer. The display command that
destroys a layer must be in the writer for that display, in order for
the displays to run independently. Do so in the test itself, rather than
relying on TearDown to destroy all of them.

Bug: 333974982
Test: atest VtsHalGraphicsComposer3_TargetTest
Flag: TEST_ONLY
Change-Id: Ib6807c85bc6cb32c56cc72126a9fa68363365912